requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-genai>=0.1.0
python-dotenv>=1.0.1
